requirements.txt
openai==1.14.1
playsound==1.3.0
PyAudio==0.2.14
PyYAML==6.0.1
SpeechRecognition==3.10.1
pocketsphinx
spotipy
flask_socketio
python-socketio[client]

#sudo apt-get install libasound-dev portaudio19-dev libportaudio2 libportaudiocpp0
#sudo apt-get install ffmpeg libav-tools
#sudo pip install pyaudio
